+++ FindPOLine/constants.js
@@ -10,7 +10,7 @@ import {
 export const FILTERS = {
   ACCESS_PROVIDER: 'accessProvider',
   ACQUISITION_METHOD: 'acquisitionMethod',
-  ACQUISITIONS_UNIT: 'acqUnitIds',
+  ACQUISITIONS_UNIT: 'purchaseOrder.acqUnitIds',
   ACTIVATED: 'activated',
   ACTUAL_RECEIPT_DATE: 'receiptDate',
   CLAIM_GRACE: 'grace',
@@ -28,17 +28,17 @@ export const FILTERS = {
   MATERIAL_TYPE_PHYSICAL: 'materialTypePhysical',
   ORDER_FORMAT: 'orderFormat',
   PAYMENT_STATUS: 'paymentStatus',
-  PREFIX: 'poNumberPrefix',
+  PREFIX: 'purchaseOrder.poNumberPrefix',
   RECEIPT_DUE: 'receiptDue',
   RECEIPT_STATUS: 'receiptStatus',
   RUSH: 'rush',
   SOURCE_CODE: 'source',
   SUBSCRIPTION_FROM: 'subscriptionFrom',
   SUBSCRIPTION_TO: 'subscriptionTo',
-  SUFFIX: 'poNumberSuffix',
+  SUFFIX: 'purchaseOrder.poNumberSuffix',
   TAGS: 'tags',
   TRIAL: 'trial',
-  VENDOR: 'vendor',
+  VENDOR: 'purchaseOrder.vendor',
 };
 
 export const ACQUISITION_METHOD_FILTER_OPTIONS = Object.keys(ACQUISITION_METHOD).map(key => ({